+++ src/mapchete_hub_cli/__init__.py
@@ -2,4 +2,5 @@ from mapchete_hub_cli._client import Client, COMMANDS, DEFAULT_TIMEOUT, JOB_STAT
 
 
 __all__ = ["Client", "COMMANDS", "DEFAULT_TIMEOUT", "JOB_STATES"]
+
 __version__ = "2023.6.0"